test(memory): sortir les tests de l'embedder sous le plafond de 500 NLOC
Le correctif de resilience a porte embedder.rs de 378 a 560 lignes, franchissant
le plafond de 500 NLOC que Codacy applique de facon bloquante sur ce depot. Les
autres fichiers touches depassaient deja avant ce lot ; celui-ci est le seul que
la PR fait basculer.

Les tests partent dans embedder_tests.rs via le #[path] deja utilise par
clock.rs, schema.rs et dated_context.rs. Aucun changement de comportement.
